LayoutTests/http/tests/preload/picture-type-expected.txt

PASS internals.isPreloaded('/resources/square20.jpg?0&preload'); is true
PASS internals.isPreloaded('/resources/square20.jpg?0'); is false
PASS internals.isPreloaded('/resources/square20.jpg?1&preload'); is true
PASS internals.isPreloaded('/resources/square20.jpg?1'); is false
PASS internals.isPreloaded('/resources/square20.jpg?2&preload'); is true
PASS internals.isPreloaded('/resources/square20.jpg?2'); is false
PASS internals.isPreloaded('/resources/square20.jpg?3&preload'); is true
PASS internals.isPreloaded('/resources/square20.jpg?3'); is false
PASS internals.isPreloaded('/resources/square20.jpg?4&preload'); is true
PASS internals.isPreloaded('/resources/square20.jpg?4&100w'); is false
PASS internals.isPreloaded('/resources/square20.jpg?4&1000w'); is false
PASS internals.isPreloaded('/resources/square20.jpg?4'); is false
PASS internals.isPreloaded('/resources/square20.jpg?5&preload'); is true
PASS internals.isPreloaded('/resources/square20.jpg?5&100w'); is false
PASS internals.isPreloaded('/resources/square20.jpg?5&800w'); is false
PASS internals.isPreloaded('/resources/square20.jpg?5&1000w'); is false
PASS internals.isPreloaded('/resources/square20.jpg?6&preload'); is true
PASS internals.isPreloaded('/resources/square20.jpg?6'); is false
PASS internals.isPreloaded('/resources/square20.jpg?7&preload'); is true
PASS internals.isPreloaded('/resources/square20.jpg?7'); is false
PASS internals.isPreloaded('/resources/square20.jpg?8&preload'); is true
PASS internals.isPreloaded('/resources/square20.jpg?8'); is false
PASS internals.isPreloaded('/resources/square20.jpg?9&preload'); is true
PASS internals.isPreloaded('/resources/square20.jpg?9'); is false
PASS internals.isPreloaded('/resources/square20.jpg?10&preload'); is true
PASS internals.isPreloaded('/resources/square20.jpg?10&dup'); is false
PASS internals.isPreloaded('/resources/square20.jpg?10'); is false
PASS internals.isPreloaded('/resources/square20.jpg?11&preload'); is true
PASS internals.isPreloaded('/resources/square20.jpg?11&bad'); is false
PASS internals.isPreloaded('/resources/square20.jpg?11'); is false
PASS internals.isPreloaded('/resources/square20.jpg?12&preload'); is true
PASS internals.isPreloaded('/resources/square20.jpg?12&bad'); is false
PASS internals.isPreloaded('/resources/square20.jpg?12'); is false
PASS internals.isPreloaded('/resources/square20.jpg?13&preload'); is true
PASS internals.isPreloaded('/resources/square20.jpg?13&bad'); is false
PASS internals.isPreloaded('/resources/square20.jpg?13&invalid'); is false
PASS successfullyParsed is true

TEST COMPLETE
